test: cover parse_ssml edge cases
